fix: MessageBubbleのディレクトリ変更
